Update main.go to import subcommands
Our new subcommands are invisible because we aren't importing them
anywhere. In this commit, we import them into the top-level `main.go` file.
Because we're only interested in the side-effects of the import (including
the packages in our build so that the `init` functions will run), we need
to use the blank identifier (`_`) for the imports. From [1]:

    To import a package solely for its side-effects (initialization), use
    the blank identifier as explicit package name:

    import _ "lib/math"

[1]: https://go.dev/ref/spec#Import_declarations